special functions hierarchy renamed and documented
