+++ plugins/formal-verification/formver.core/src/org/jetbrains/kotlin/formver/embeddings/callables/SpecialKotlinFunction.kt
@@ -1,32 +1,23 @@
 /*
- * Copyright 2010-2023 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Copyright 2010-2024 JetBrains s.r.o. and Kotlin Programming Language contributors.
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
 package org.jetbrains.kotlin.formver.embeddings.callables
 
-import org.jetbrains.kotlin.formver.embeddings.expression.*
-import org.jetbrains.kotlin.formver.embeddings.types.buildFunctionPretype
-import org.jetbrains.kotlin.formver.names.*
-import org.jetbrains.kotlin.formver.names.NameMatcher
+import org.jetbrains.kotlin.formver.conversion.StmtConversionContext
+import org.jetbrains.kotlin.formver.embeddings.expression.ExpEmbedding
 import org.jetbrains.kotlin.formver.viper.ast.Method
-import org.jetbrains.kotlin.name.CallableId
-import org.jetbrains.kotlin.name.FqName
-import org.jetbrains.kotlin.name.Name
-import org.jetbrains.kotlin.formver.embeddings.expression.OperatorExpEmbeddings.AddIntInt
-import org.jetbrains.kotlin.formver.embeddings.expression.OperatorExpEmbeddings.AddCharInt
-import org.jetbrains.kotlin.formver.embeddings.expression.OperatorExpEmbeddings.AddStringString
-import org.jetbrains.kotlin.formver.embeddings.expression.OperatorExpEmbeddings.DivIntInt
-import org.jetbrains.kotlin.formver.embeddings.expression.OperatorExpEmbeddings.MulIntInt
-import org.jetbrains.kotlin.formver.embeddings.expression.OperatorExpEmbeddings.Not
-import org.jetbrains.kotlin.formver.embeddings.expression.OperatorExpEmbeddings.StringGet
-import org.jetbrains.kotlin.formver.embeddings.expression.OperatorExpEmbeddings.SubCharChar
-import org.jetbrains.kotlin.formver.embeddings.expression.OperatorExpEmbeddings.SubCharInt
-import org.jetbrains.kotlin.formver.embeddings.expression.OperatorExpEmbeddings.SubIntInt
-import org.jetbrains.kotlin.formver.viper.MangledName
 
-
-interface SeparatelyHandledKotlinFunction : FunctionEmbedding {
+/**
+ * Some functions are handled specially by our conversion.
+ * This can be helpful when proving. For example, `Int.plus(Int)` will
+ * eventually be translated to native operator `+` in Viper.
+ *
+ * In particular, that means that there won't be a call to the Viper method
+ * sometimes.
+ */
+interface SpecialKotlinFunction : FunctionEmbedding {
     val packageName: List<String>
     val className: String?
         get() = null
@@ -34,149 +25,31 @@ interface SeparatelyHandledKotlinFunction : FunctionEmbedding {
 }
 
 /**
- * Kotlin function that should be handled specially by our conversion.
- *
- * This includes `contract` and operations on primitive types, where providing a full embedding into Viper
- * offers more possibilities for reasoning about the code.
+ * Kotlin function that will always be handled specially, like aforementioned `Int.plus(Int)`.
  */
-interface SpecialKotlinFunction : SeparatelyHandledKotlinFunction {
+interface FullySpecialKotlinFunction : SpecialKotlinFunction {
     override val viperMethod: Method?
         get() = null
 }
 
-val SeparatelyHandledKotlinFunction.callableId: CallableId
-    get() = CallableId(FqName.fromSegments(packageName), className?.let { FqName(it) }, Name.identifier(name))
-
-fun SeparatelyHandledKotlinFunction.embedName(): ScopedKotlinName = callableId.embedFunctionName(callableType)
-
-object SpecialKotlinFunctions {
-    private val contractBuilderTypeName = buildName {
-        packageScope(listOf("kotlin", "contracts"))
-        ClassKotlinName(listOf("ContractBuilder"))
+/**
+ * Kotlin function that will sometimes be handled specially depending on arguments they're called with.
+ *
+ * This is useful in cases like `String.plus(Any)`. If `Any` turns out to be a `String` we can substitute
+ * sequence concatenation in Viper. Otherwise, we're forced to call an actual method (which is stored in `baseEmbedding`).
+ *
+ * Currently, `String.plus(Any)` is the sole case when we use this interface.
+ */
+interface PartiallySpecialKotlinFunction : SpecialKotlinFunction {
+    val baseEmbedding: FunctionEmbedding?
+    fun tryInsertCall(args: List<ExpEmbedding>, ctx: StmtConversionContext): ExpEmbedding?
+    override fun insertCall(args: List<ExpEmbedding>, ctx: StmtConversionContext): ExpEmbedding {
+        return tryInsertCall(args, ctx) ?: baseEmbedding?.insertCall(args, ctx)
+        ?: error("Base embedding for partially special function $name not specified")
     }
-    private val booleanArrayTypeName = buildName {
-        packageScope(listOf("kotlin"))
-        ClassKotlinName(listOf("BooleanArray"))
-    }
 
-    val byName: Map<MangledName, FunctionEmbedding> = buildSpecialFunctions {
-        val intIntToIntType = buildFunctionPretype {
-            withDispatchReceiver { int() }
-            withParam { int() }
-            withReturnType { int() }
-        }
-        withCallableType(intIntToIntType) {
-            addFunction("kotlin", className = "Int", name = "plus") { args, _ ->
-                AddIntInt(args[0], args[1])
-            }
-            addFunction("kotlin", className = "Int", name = "minus") { args, _ ->
-                SubIntInt(args[0], args[1])
-            }
-            addFunction("kotlin", className = "Int", name = "times") { args, _ ->
-                MulIntInt(args[0], args[1])
-            }
-            addFunction("kotlin", className = "Int", name = "div") { args, _ ->
-                blockOf(
-                    InhaleDirect(NeCmp(args[1], IntLit(0))),
-                    DivIntInt(args[0], args[1]),
-                )
-            }
-        }
+    fun initBaseEmbedding(embedding: FunctionEmbedding)
 
-        val booleanToBooleanType = buildFunctionPretype {
-            withDispatchReceiver { boolean() }
-            withReturnType { boolean() }
-        }
-
-        addFunction(booleanToBooleanType, "kotlin", className = "Boolean", name = "not") { args, _ ->
-            Not(args[0])
-        }
-
-        val verifyCallableType = buildFunctionPretype {
-            withParam {
-                klass {
-                    withName(booleanArrayTypeName)
-                }
-            }
-            withReturnType { unit() }
-        }
-        addFunction(verifyCallableType, "org", "jetbrains", "kotlin", "formver", "plugin", name = "verify") { args, _ ->
-            args.map { Assert(it) }.toBlock()
-        }
-
-        val contractCallableType = buildFunctionPretype {
-            withParam {
-                function {
-                    withDispatchReceiver {
-                        klass {
-                            withName(contractBuilderTypeName)
-                        }
-                    }
-                    withReturnType { unit() }
-                }
-            }
-            withReturnType { unit() }
-        }
-
-        addFunction(contractCallableType, "kotlin", "contracts", name = "contract") { _, _ ->
-            UnitLit
-        }
-
-        val charCharToIntType = buildFunctionPretype {
-            withDispatchReceiver { char() }
-            withParam { char() }
-            withReturnType { int() }
-        }
-
-        addFunction(charCharToIntType, "kotlin", className = "Char", name = "minus") { args, _ ->
-            SubCharChar(args[0], args[1])
-        }
-
-        val charIntToCharType = buildFunctionPretype {
-            withDispatchReceiver { char() }
-            withParam { int() }
-            withReturnType { char() }
-        }
-
-        withCallableType(charIntToCharType) {
-            addFunction("kotlin", className = "Char", name = "plus") { args, _ ->
-                AddCharInt(args[0], args[1])
-            }
-            addFunction("kotlin", className = "Char", name = "minus") { args, _ ->
-                SubCharInt(args[0], args[1])
-            }
-        }
-
-        val stringIntToCharType = buildFunctionPretype {
-            withDispatchReceiver { string() }
-            withParam { int() }
-            withReturnType { char() }
-        }
-
-        addFunction(stringIntToCharType, "kotlin", className = "String", name = "get") { args, _ ->
-            StringGet(args[0], args[1])
-        }
-
-        val stringStringToStringType = buildFunctionPretype {
-            withDispatchReceiver { string() }
-            withParam { string() }
-            withReturnType { string() }
-        }
-
-        addFunction(stringStringToStringType, "kotlin", className = "String", name = "plus") { args, _ ->
-            AddStringString(args[0], args[1])
-        }
-    }
+    override val viperMethod: Method?
+        get() = baseEmbedding?.viperMethod
 }
-
-val FunctionEmbedding.isVerifyFunction: Boolean
-    get() = this is SpecialKotlinFunction && NameMatcher.matchClassScope(this.embedName()) {
-        ifPackageName("org", "jetbrains", "kotlin", "formver", "plugin") {
-            ifNoReceiver {
-                ifFunctionName("verify") {
-                    return true
-                }
-            }
-        }
-        return false
-    }
